fix quoting in tools run script
The run script used for tools accepts a list of arguments for a command it is intended to wrap. It needs to be able to run this directly, as well as pass it to weston-terminal as a single string.

The previous code attempts this by attempting multiple ways of passing the arguments, and retrying a different way on error. This could have unexpected results, and results in ugly messages shown to users.

Various attempts at this quoting have been done in the past, but the correct way to quote commands to use directly in bash as an array is "$@". Creating the single string for weston-terminal is not as trivial, but printf includes a %q format specification that will quote things like we need.